Fixed three compilation errors in 64-bit Ubuntu
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@287 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/delilah/Workspace.cpp
@@ -106,7 +106,7 @@ void Workspace::updateJob(unsigned int id, bool finished, bool error, QString me
 /*
  * SLOT.
  */
-void Workspace::finishJob(size_t id, bool error, QString message)
+void Workspace::finishJob(unsigned int id, bool error, QString message)
 {
 	// TODO:
 	int index = findJobIndex(id);
@@ -162,7 +162,7 @@ void Workspace::finishJob(size_t id, bool error, QString message)
 }
 
 
-int Workspace::findJobIndex(size_t id)
+int Workspace::findJobIndex(unsigned int id)
 {
 	for(int i=0; i<jobs.size(); i++)
 	{
